+++ data-service-jvfs/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.4 - 2023-11-12]
+
+### Chanted
+
+- Added generation id : ${yyyyMMddHHmmss}_${UUID}
+
 ## [0.2.3 - 2023-11-12]
 
 ### Added